Revert juce code to a known working version

+++ libs/juce/source/modules/juce_audio_basics/effects/juce_Reverb.h
@@ -333,7 +333,7 @@ private:
                 if (countdown <= 0)
                     currentValue = target;
                 else
-                    step = (target - currentValue) / (float) countdown;
+                    step = (target - currentValue) / countdown;
             }
         }
