+++ libs/juce/source/modules/juce_core/files/juce_FileInputStream.cpp
@@ -57,7 +57,7 @@ int FileInputStream::read (void* buffer, int bytesToRead)
     jassert (buffer != nullptr && bytesToRead >= 0);
 
     const size_t num = readInternal (buffer, (size_t) bytesToRead);
-    currentPosition += (int64) num;
+    currentPosition += num;
 
     return (int) num;
 }
